Text Analysis and Convolutional Neural Networks
In the first part of my final assignment for Machine Learning, I work with data which consists of headlines that signify clickbait or not. I test out three classification models with different tokenizers - logistic regression, logistic regression with bi-gram vectorization, and logistic regression with lemmatization tokenizer - and determine which model is the best for determining which headlines are clickbait.

In part two, I build a convolutional neural network with 2 hidden layers on the IRIS dataset, and find the optimal number of hidden nodes for each unit using gridsearch.